requirements.txt
discord.py
aiogTTS
akinator.py
asyncdagpi
asyncpg
git+https://github.com/Rapptz/discord-ext-menus@master
doc-search
humanize
mystbin.py
git+https://github.com/ShineyDev/prettify.py@master
psutil
sr-api
tag-formatter
toml
topggpy
https://github.com/edugomez102/twemoji-parser
